feat(planner): unify read semantics across vtable and file operations
- Introduced a comprehensive plan to consolidate read operations in `vtable_read.rs`, eliminating duplicate logic in `lix_state`, `lix_state_by_version`, and `entity_views/read`.
- Established a new internal reader/writer interface for file operations, merging `lix_file`, `lix_file_by_version`, and `lix_file_history` into a single semantic planner.
- Created a structured migration plan with defined phases to ensure smooth transition and maintain performance benchmarks.
- Implemented initial phases of the vtable read planner, including the introduction of `VtableReadOp` and `VtableReadPlan` for normalized intent and semantic planning.
- Captured baseline performance metrics for existing read operations to measure improvements post-implementation.
