+++ packages/engine/src/filesystem/mod.rs
@@ -1,4 +1,5 @@
 pub(crate) mod mutation_rewrite;
 pub(crate) mod path;
 pub(crate) mod pending_file_writes;
+pub(crate) mod planner;
 pub(crate) mod select_rewrite;
